fix(library): address review comments on dtype and contract examples
- cast labels to int64 before postprocessing, cast label_ids back to
  int32 to match the Prediction contract
- reword test_model_contract.py docstring; samples_to_tensors() accepts
  Collatable, but public predict()/fit() stay Sample|list[Sample]|Batch
- convert SAM mask output to numpy in README example

+++ library/examples/gradio_postprocessing.py
@@ -156,7 +156,7 @@ def postprocess(prediction: Prediction, pipeline: PostProcessor | None, device: 
 
     masks = torch.as_tensor(prediction.masks, device=device).bool()
     scores = torch.as_tensor(prediction.scores, device=device).float()
-    labels = torch.as_tensor(prediction.label_ids, device=device)
+    labels = torch.as_tensor(prediction.label_ids, device=device).long()
 
     new_masks, new_scores, new_labels = pipeline(masks, scores, labels)
 
@@ -164,7 +164,7 @@ def postprocess(prediction: Prediction, pipeline: PostProcessor | None, device: 
     if new_masks.shape[0] > 0:
         boxes = masks_to_boxes_traceable(new_masks).cpu().numpy()
 
-    label_ids = new_labels.cpu().numpy()
+    label_ids = new_labels.cpu().numpy().astype(np.int32)
     return Prediction(
         masks=new_masks.cpu().numpy(),
         scores=new_scores.cpu().numpy(),
